+++ src/controllers/userController.js
@@ -7,8 +7,9 @@ const asyncHandler = require("../middlewares/asyncHandler");
 const {uploadSingleImage} = require("../middlewares/uploadImageMiddleware");
 const apiSuccess = require("../utils/apiSuccess");
 const ApiError = require("../utils/apiError");
-const UserModel = require("../models/userModel");
 const generateJWT = require("../utils/generateJWT");
+const getImageUrl = require("../utils/getImageUrl");
+const UserModel = require("../models/userModel");
 
 const getAllUsers = asyncHandler(async (req, res, next) => {
     const usersCount = await UserModel.countDocuments();
@@ -34,25 +35,22 @@ const getAllUsers = asyncHandler(async (req, res, next) => {
         sortBy = req.query.sort.split(',').join(" ");
     }
 
-    let limitField = "-__v -password";
-    if (req.query.fields) {
-        limitField = req.query.fields.split(",").join(" ");
-    }
+    const selectedField = "name email phone profileImg addresses gender accountActive";
 
     const users = await UserModel
         .find()
         .limit(limit)
         .skip(skip)
         .sort(sortBy)
-        .select(limitField);
+        .select(selectedField);
 
     if (!users) {
-        return next(new ApiError(`No users found`, 404));
+        return next(new ApiError(`no users found`, 404));
     }
 
     return res.status(200).json(
         apiSuccess(
-            `users Found`,
+            `users found`,
             200,
             {
                 pagination,
@@ -64,52 +62,22 @@ const getAllUsers = asyncHandler(async (req, res, next) => {
 const getUser = asyncHandler(async (req, res, next) => {
     const {id} = req.params;
 
-    const user = await UserModel.findById(id, "-password -__v");
+    const selectedField = "name email phone profileImg addresses gender accountActive";
+
+    const user = await UserModel.findById(id, selectedField);
 
     if (!user) {
         return next(new ApiError(`no user for this id ${id}`, 404))
     }
 
+    user.profileImg = getImageUrl(req, user.profileImg);
+
     return res.status(200).json(
         apiSuccess(
             "user found successfully",
             200,
-            {user}
+            user
         ));
-});
-
-const uploadProfileImage = uploadSingleImage("profileImg");
-
-const resizeProfileImage = asyncHandler(async (req, res, next) => {
-    const fileName = `user-${Math.round(
-        Math.random() * 1e9
-    )}-${Date.now()}.jpeg`;
-
-    if (req.file) {
-        await sharp(req.file.buffer)
-            .resize(600, 600)
-            .toFormat("jpeg")
-            .jpeg({quality: 95})
-            .toFile(`uploads/users/${fileName}`)
-            .then(async () => {
-                const user = await UserModel.findById(req.params.id);
-                const oldFileName = user.profileImg;
-                const filePath = `uploads/users/${oldFileName}`;
-                fs.access(filePath, fs.constants.F_OK, (err) => {
-                    if (!err) {
-                        // File exists, so delete it
-                        fs.unlink(filePath, (deleteErr) => {
-                            if (deleteErr) {
-                                console.error('Error deleting file:', deleteErr);
-                            }
-                        });
-                    }
-                });
-            });
-
-        req.body.profileImg = fileName;
-    }
-    next();
 });
 
 const updateUser = asyncHandler(async (req, res) => {
@@ -120,7 +88,6 @@ const updateUser = asyncHandler(async (req, res) => {
         phone: req.body.phone,
         address: req.body.address,
         gender: req.body.gender,
-        profileImg: req.body.profileImg,
     }, {
         new: true,
     });
@@ -132,6 +99,52 @@ const updateUser = asyncHandler(async (req, res) => {
             null
         ));
 
+});
+
+const uploadProfileImage = uploadSingleImage("profileImg");
+
+const resizeProfileImage = asyncHandler(async (req, res, next) => {
+    const fileName = `user-${Math.round(
+        Math.random() * 1e9
+    )}-${Date.now()}.jpeg`;
+
+    if (req.file) {
+        await sharp(req.file.buffer)
+            .toFormat("jpeg")
+            .jpeg({quality: 95})
+            .toFile(`uploads/users/${fileName}`)
+            .then(async () => {
+                const user = await UserModel.findById(req.params.id);
+                const oldFileName = user.profileImg;
+                const filePath = `uploads/users/${oldFileName}`;
+
+                if (user.profileImg !== "defaultImage.png") {
+                    fs.access(filePath, fs.constants.F_OK, (err) => {
+                        if (!err) {
+                            fs.unlink(filePath, (deleteErr) => {
+                                if (deleteErr) {
+                                    return next(new ApiError(`something happen when change profileImg`, 400));
+                                }
+                            });
+                        }
+                    });
+                }
+            });
+
+        req.body.profileImg = fileName;
+    }
+    next();
+});
+
+const updateProfileImage = asyncHandler(async (req, res) => {
+    await UserModel.findById(req.loggedUser._id, {profileImg: req.body.profileImg});
+
+    return res.status(200).json(
+        apiSuccess(
+            "profile image updated successfully",
+            200,
+            null,
+        ));
 });
 
 const search = asyncHandler(async (req, res, next) => {
@@ -146,7 +159,9 @@ const search = asyncHandler(async (req, res, next) => {
         {address: {$regex: keyword, $options: "i"},}
     ]
 
-    const users = await UserModel.find(queryObj, "-password -__v");
+    const selectedField = "name email phone profileImg addresses gender accountActive";
+
+    const users = await UserModel.find(queryObj, selectedField);
 
     if (!users) {
         return next(new ApiError(`No users found matched this search key: ${keyword}`, 404));
@@ -156,7 +171,7 @@ const search = asyncHandler(async (req, res, next) => {
         apiSuccess(
             `users Found`,
             200,
-            {users}
+            users
         ));
 });
 
@@ -167,16 +182,19 @@ const deleteUser = asyncHandler(async (req, res) => {
 
     const oldFileName = user.profileImg;
     const filePath = `uploads/users/${oldFileName}`;
-    fs.access(filePath, fs.constants.F_OK, (err) => {
-        if (!err) {
-            // File exists, so delete it
-            fs.unlink(filePath, (deleteErr) => {
-                if (deleteErr) {
-                    console.error('Error deleting file:', deleteErr);
-                }
-            });
-        }
-    });
+
+    if (user.profileImg !== "defaultImage.png") {
+        fs.access(filePath, fs.constants.F_OK, (err) => {
+            if (!err) {
+                // File exists, so delete it
+                fs.unlink(filePath, (deleteErr) => {
+                    if (deleteErr) {
+                        console.error('Error deleting file:', deleteErr);
+                    }
+                });
+            }
+        });
+    }
 
     return res.status(200).json(
         apiSuccess(
@@ -187,7 +205,7 @@ const deleteUser = asyncHandler(async (req, res) => {
 
 });
 
-const getUserProfile = asyncHandler(async (req, res, next) => {
+const setProfileID = asyncHandler(async (req, res, next) => {
     req.params.id = req.loggedUser._id;
     next();
 });
@@ -215,27 +233,6 @@ const changePassword = asyncHandler(async (req, res) => {
             {token}
         ));
 })
-
-const updateProfile = asyncHandler(async (req, res) => {
-    const id = req.loggedUser._id;
-
-    await UserModel.findByIdAndUpdate(id, {
-        name: req.body.name,
-        phone: req.body.phone,
-        address: req.body.address,
-        profileImg: req.body.profileImg,
-    }, {
-        new: true,
-    });
-
-    return res.status(200).json(
-        apiSuccess(
-            `user updated successfully`,
-            200,
-            null
-        ));
-
-});
 
 const addUserAddress = asyncHandler(async (req, res) => {
     const user = await UserModel.findById(req.loggedUser._id);
@@ -270,10 +267,12 @@ const addUserAddress = asyncHandler(async (req, res) => {
 });
 
 const removeUserAddress = asyncHandler(async (req, res) => {
+    const {addressId} = req.params;
+
     const user = await UserModel.findByIdAndUpdate(
         req.loggedUser._id,
         {
-            $pull: {addresses: {_id: req.params.addressId}},
+            $pull: {addresses: {_id: addressId}},
         },
         {new: true}
     );
@@ -282,7 +281,7 @@ const removeUserAddress = asyncHandler(async (req, res) => {
         apiSuccess(
             "Address removed successfully",
             200,
-            user.addresses
+            {address: user.addresses}
         ));
 });
 
@@ -293,7 +292,7 @@ const getProfileAddresses = asyncHandler(async (req, res) => {
         apiSuccess(
             "Address Founded successfully",
             200,
-            user.addresses
+            {address: user.addresses}
         ));
 });
 
@@ -306,9 +305,9 @@ module.exports = {
     resizeProfileImage,
     changePassword,
     search,
-    getUserProfile,
-    updateProfile,
+    setProfileID,
     addUserAddress,
     removeUserAddress,
     getProfileAddresses,
+    updateProfileImage,
 }
